+++ src/components/rental/image.svelte
@@ -1,8 +1,22 @@
 <script lang="ts">
 	export let src: string;
 	export let alt: string;
+
+	let isLarge = false;
+
+	function onToggleSize() {
+		isLarge = !isLarge;
+	}
 </script>
 
-<div class="image">
-	<img {src} {alt} />
-</div>
+{#if isLarge}
+	<button on:click={onToggleSize} class="image large">
+		<img {src} {alt} />
+		<small>View Smaller</small>
+	</button>
+{:else}
+	<button on:click={onToggleSize} class="image">
+		<img {src} {alt} />
+		<small>View Larger</small>
+	</button>
+{/if}